requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
groq==0.22.0
pypdf2==3.0.1
pillow==11.2.1
matplotlib==3.10.1
seaborn==0.13.2
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0